+++ mycelium/src/router.rs
@@ -260,6 +260,52 @@ where
         }
     }
 
+    /// Get a [`SharedSecret`] for a remote, if a selected route exists to the remote.
+    // TODO: Naming
+    pub fn get_shared_secret_if_selected(&self, dest: IpAddr) -> Option<SharedSecret> {
+        // TODO: proper fix
+        match self.routing_table.best_routes(dest) {
+            Routes::Exist(routes) => {
+                if routes.selected().is_some() {
+                    Some(routes.shared_secret().clone())
+                } else {
+                    // Optimistically try to fetch a new route for the subnet for next time.
+                    // TODO: this can likely be handled better, but that relies on continuously
+                    // quering routes in use, and handling unfeasible routes
+                    if let Some(route_entry) = routes.iter().next() {
+                        // We have a fallback route, use the source key from that to do a seqno
+                        // request. Since the next hop might be dead, just do a broadcast. This
+                        // might be blocked by the seqno request cache.
+                        self.send_seqno_request(route_entry.source(), None, None);
+                    } else {
+                        // We don't have any routes, so send a route request. this might fail due
+                        // to the source table.
+                        self.send_route_request(
+                            Subnet::new(dest, 64)
+                                .expect("64 is a valid subnet size for an IPv6 address; qed"),
+                        );
+                    }
+
+                    None
+                }
+            }
+            Routes::Queried => {
+                tokio::task::block_in_place(|| std::thread::sleep(QUERY_CHECK_DURATION));
+                self.get_shared_secret_if_selected(dest)
+            }
+            Routes::NoRoute => None,
+            Routes::None => {
+                // NOTE: we request the full /64 subnet
+                self.send_route_request(
+                    Subnet::new(dest, 64)
+                        .expect("64 is a valid subnet size for an IPv6 address; qed"),
+                );
+                tokio::task::block_in_place(|| std::thread::sleep(QUERY_CHECK_DURATION));
+                self.get_shared_secret_if_selected(dest)
+            }
+        }
+    }
+
     /// Gets the cached [`SharedSecret`] based on the associated [`PublicKey`] of the remote.
     #[inline]
     pub fn get_shared_secret_by_pubkey(&self, dest: &PublicKey) -> Option<SharedSecret> {